feat: reorder expressions to handle operator precedence (#3)
* refactor: add partial tags to printer

I would be interested in turning this into a full valid xml printer in the future...

Just added this for now to help with some of the debugging

* chore: add rough input file

* Add operator precedence handler WIP

* chore: remove TODO comment

* chore: add precedence handler to index for debugging

* feat: support a++ + b expressions

* fix: missing precedence for null coalescing

* feat: reorder based on operator precedence

* test: add output for precedence handler

* test: add missing prec handler to existing tests

Do we really want to be manually instantiating an entirely new instance of each visitor for each test suite? What about providing a default instantiation in testParsing?

* test: add basic test for expression reordering

* revert: "refactor: add partial tags to printer"

This reverts commit 0f9f581393d11046ddea61b00d7a044b86defb5f.

We want to add a more fully fledged feature for this later.