chore: migrate to monorepo structure (#4)
* chore: add deno

* chore: add biome

* chore: remove eslint

* chore: remove prettier

* chore: remove npm

* chore: remove tsconfig

* chore: align deno formatter with biome

* chore: prepare deno config for new workspaces

* refactor: port parser over to deno

Copy over files from og parser project, rewriting as needed in a more Deno-idiomatic manner

* refactor: port generator over to deno

Combine diagram & type declaration generators into one interface and rewrite in a more Deno-idiomatic manner

* chore: add generate command

* chore: add deno.lock

* chore: generate CST syntax diagrams & type declarations

* feat: reorder expressions to handle operator precedence (#3)

* refactor: add partial tags to printer

I would be interested in turning this into a full valid xml printer in the future...

Just added this for now to help with some of the debugging

* chore: add rough input file

* Add operator precedence handler WIP

* chore: remove TODO comment

* chore: add precedence handler to index for debugging

* feat: support a++ + b expressions

* fix: missing precedence for null coalescing

* feat: reorder based on operator precedence

* test: add output for precedence handler

* test: add missing prec handler to existing tests

Do we really want to be manually instantiating an entirely new instance of each visitor for each test suite? What about providing a default instantiation in testParsing?

* test: add basic test for expression reordering

* revert: "refactor: add partial tags to printer"

This reverts commit 0f9f581393d11046ddea61b00d7a044b86defb5f.

We want to add a more fully fledged feature for this later.

* test: add preflight tests

* test: add test utils

* test: add end-to-end tests

* refactor: port over precedence handler cst visitor

* test: port over integration tests

* chore: update deno.lock

* fix: number literal precedence

The lexer was not recognizing a real number because it parsed it as an int and then choked on the decimal point

* chore: remove old parser project

Remove the original parser workspace, completing the move to the monorepo

* Revert "chore: remove old parser project"

This reverts commit e5663435871244a70c94ff51b67396136e78ee98.

---------

Co-authored-by: Aidan Jones <aejones4gm@gmail.com>